validations after retraining

+++ python/config.py
@@ -21,25 +21,29 @@ configuration.model.classifiers = {
     'cv-0' : os.path.join(configuration.home,"models/model-binary-cv-0"),
     'window-final': os.path.join(configuration.home,"models/model-window"),
     'window-cv-0': os.path.join(configuration.home,"models/model-convnet-0"),
+    'window-cv-5': os.path.join(configuration.home,"models/model_window_cv_5_49"),
     }
 configuration.model.classifier = configuration.model.classifiers['cv-0']
 
 configuration.model.classifier_weights = os.path.join(configuration.home,"models/model-binary.h5")
 configuration.model.classifier_mean = 127.367759008 #mean_value 
 configuration.model.classifier_max = 43.5644119735
-configuration.model.window = configuration.model.classifiers['window-cv-0']
+configuration.model.window = configuration.model.classifiers['window-cv-5']
 configuration.model.window_templates = os.path.join(configuration.home,"models/templates")
+configuration.model.min_probability = 0.80
 #configuration.model.window_mean = 166.169432427  #166.166054305 max 46.4016257135
 #configuration.model.window_max = 46.3409904943 #mean 166.169432427 max 46.3409904943 166.169432427 max 46.3409904943
 
 configuration.input = Dummy()
-configuration.input.image_with = 384
+configuration.input.image_width = 384
 configuration.input.image_height = 288
+configuration.input.binary_width = 384 / 4
+configuration.input.binary_height = 288 /4
 
 configuration.window = Dummy()
-configuration.window.image_with = 96 #96
+configuration.window.image_width = 96 #96
 configuration.window.image_height = 72 #72
-configuration.window.offset_with = 10 #10   
+configuration.window.offset_width = 10 #10   
 configuration.window.offset_height = 10 #10
 configuration.window.offset_class = 6 #classifier start in 0
 configuration.window.start = 6
@@ -54,7 +58,7 @@ configuration.roofly.template = "EXPERIMENT_T{tube}_L{window}_{date}_{time}_{ses
 def get_configuration():
     return configuration
     
-def setup_video_folders(video_filname):
+def setup_video_folders(video_filname,reset_folder=True):
     #extract video name
     drive, path = os.path.splitdrive(video_filname)
     path, video_name = os.path.split(path)
@@ -66,7 +70,8 @@ def setup_video_folders(video_filname):
     video_folder = os.path.join(configuration.home,"processing",video_name)
 
     #delete video folder
-    shutil.rmtree(video_folder,ignore_errors=True)
+    if reset_folder:
+        shutil.rmtree(video_folder,ignore_errors=True)
 
 
     if not os.path.exists(video_folder):
